cli/requirements.txt
numpy==1.12.1
tqdm==4.11.2
Pillow==5.4.1
